Turned off the selective zero truncation for the local properties...
NOTE this was really tricky to merge with the new local property stuff
in turbogap, so ended up changing the way the zero-truncation flag is
stored. Hopefully this will do exactly the same thing as before, but it
needs to be checked.

NOTE 2: turned off entirely for the GPU version, since it was
removed sometime earlier (and only needs to be re-implemented for
Hirshfeld volumes). Now setting the option just emits a warning
and continues as usual

Haista paska, Fortran